feat: created analyticsLog directory and moved files accordingly
