+++ api.planx.uk/modules/analytics/analyticsLog/controller.ts
@@ -1,6 +1,6 @@
 import { z } from "zod";
-import { trackAnalyticsLogExit } from "./service.js";
-import type { ValidatedRequestHandler } from "../../shared/middleware/validate.js";
+import { trackAnalyticsLogExit } from "../service.js";
+import type { ValidatedRequestHandler } from "../../../shared/middleware/validate.js";
 
 export const logAnalyticsSchema = z.object({
   query: z.object({
